corrected - Button to add to favorite tracks in the night theme -Moved classes to the data package
