+++ app/src/main/java/com/example/playlistmaker/di/dbModule.kt
@@ -1,7 +1,7 @@
 package com.example.playlistmaker.di
 
 import androidx.room.Room
-import com.example.playlistmaker.db.AppDatabase
+import com.example.playlistmaker.data.AppDatabase
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
